+++ src-docs/src/views/datagrid/datagrid.js
@@ -395,18 +395,18 @@ export default () => {
         : null;
       const [content, setContent] = useState(_data);
 
-      // if (columnId === 'account') {
-      //   useEffect(() => {
-      //     let swap = 0;
-      //     const interval = setInterval(() => {
-      //       setContent(swap === 0 ? <><button>Test!</button><button>Also!</button></> : _data);
-      //
-      //       swap = swap === 0 ? 1 : 0;
-      //     }, 5000);
-      //
-      //     return () => clearInterval(interval);
-      //   }, []);
-      // }
+      if (columnId === 'account') {
+        useEffect(() => {
+          let swap = 0;
+          const interval = setInterval(() => {
+            setContent(swap === 0 ? <><button>Test!</button><button>Also!</button></> : _data);
+
+            swap = swap === 0 ? 1 : 0;
+          }, 5000);
+
+          return () => clearInterval(interval);
+        }, []);
+      }
 
       return content;
     };
